Added logger and global exception handling

+++ .eslintrc.js
@@ -1,7 +1,7 @@
 module.exports = {
     "extends": ["angular", "eslint:recommended"],
     "rules": {
-        "quotes": ["error", "double", {"avoidEscape" : true}],
+        "quotes": ["error", "single", {"avoidEscape" : true}],
         "semi": ["error", "always"]
     }
 };
